Extract progress indicator to new file.
